Restructured test to reflect library code.
