+++ test/RService.IO.Tests/Abstractions/RouteAttributeTests.cs
@@ -5,7 +5,7 @@ using FluentAssertions;
 using RService.IO.Abstractions;
 using Xunit;
 
-namespace RService.IO.Tests
+namespace RService.IO.Tests.Abstractions
 {
     public class RouteAttributeTests
     {
